// types/types.ts
export type GalleryImageDTO = {
  id: string;
  schoolId: string;
  title: string;
  description?: string;
  image: string;
  date?: string;
  categoryId: string;
  createdAt: string;
  updatedAt: string;
};
export interface GalleryImageCreateDTO {
  schoolId: string;
  title: string;
  description?: string;
  image: string;
  date?: string;
  categoryId: string;
}
export interface GalleryCategory {
  id: string;
  name: string;
}

export type Section = {
  id: string;
  schoolId: string;
  type: string;
  isComplete: boolean;
  title: string;
  subtitle?: string | null;
  isActive: boolean;
  order: number;
  settings: any;
  createdAt: Date;
  updatedAt: Date;
};